+++ log/log.go
@@ -22,19 +22,30 @@ func Init(level string, json bool) {
 	global = New(level, json)
 }
 
+// Trace prints message in trace mode.
+func Trace(msg Message) {
+	global.printf(LevelTrace, msg, os.Stdout)
+}
+
 // Debug prints message in debug mode.
 func Debug(msg Message) {
-	global.printf(levelDebug, msg, os.Stdout)
+	global.printf(LevelDebug, msg, os.Stdout)
 }
 
 // Info prints message in info mode.
 func Info(msg Message) {
-	global.printf(levelInfo, msg, os.Stdout)
+	global.printf(LevelInfo, msg, os.Stdout)
+}
+
+// Stat prints stat message regardless of the log level with info print formatting.
+// It uses printfHelper instead of printf to ignore the log level condition.
+func Stat(msg Message) {
+	global.printfHelper(LevelInfo, msg, os.Stdout)
 }
 
 // Error prints message in error mode.
 func Error(msg Message) {
-	global.printf(levelError, msg, os.Stderr)
+	global.printf(LevelError, msg, os.Stderr)
 }
 
 // Close closes logger and its channel.
@@ -47,12 +58,12 @@ func Close() {
 type Logger struct {
 	donech chan struct{}
 	json   bool
-	level  logLevel
+	level  LogLevel
 }
 
 // New creates new logger.
 func New(level string, json bool) *Logger {
-	logLevel := levelFromString(level)
+	logLevel := LevelFromString(level)
 	logger := &Logger{
 		donech: make(chan struct{}),
 		json:   json,
@@ -63,11 +74,14 @@ func New(level string, json bool) *Logger {
 }
 
 // printf prints message according to the given level, message and std mode.
-func (l *Logger) printf(level logLevel, message Message, std *os.File) {
+func (l *Logger) printf(level LogLevel, message Message, std *os.File) {
 	if level < l.level {
 		return
 	}
+	l.printfHelper(level, message, std)
+}
 
+func (l *Logger) printfHelper(level LogLevel, message Message, std *os.File) {
 	if l.json {
 		outputCh <- output{
 			message: message.JSON(),
@@ -90,40 +104,49 @@ func (l *Logger) out() {
 	}
 }
 
-// logLevel is the level of Logger.
-type logLevel int
+// LogLevel is the level of Logger.
+type LogLevel int
 
 const (
-	levelDebug logLevel = iota
-	levelInfo
-	levelError
+	LevelTrace LogLevel = iota
+	LevelDebug
+	LevelInfo
+	LevelError
 )
 
 // String returns the string representation of logLevel.
-func (l logLevel) String() string {
+func (l LogLevel) String() string {
 	switch l {
-	case levelInfo:
+	case LevelInfo:
 		return ""
-	case levelError:
+	case LevelError:
 		return "ERROR "
-	case levelDebug:
+	case LevelDebug:
 		return "DEBUG "
+	case LevelTrace:
+		// levelTrace is used for printing aws sdk logs and
+		// aws-sdk-go already adds "DEBUG" prefix to logs.
+		// So do not add another prefix to log which makes it
+		// look weird.
+		return ""
 	default:
 		return "UNKNOWN "
 	}
 }
 
-// levelFromString returns logLevel for given string. It
+// LevelFromString returns logLevel for given string. It
 // return `levelInfo` as a default.
-func levelFromString(s string) logLevel {
+func LevelFromString(s string) LogLevel {
 	switch s {
 	case "debug":
-		return levelDebug
+		return LevelDebug
 	case "info":
-		return levelInfo
+		return LevelInfo
 	case "error":
-		return levelError
+		return LevelError
+	case "trace":
+		return LevelTrace
 	default:
-		return levelInfo
+		return LevelInfo
 	}
 }